libnm: deprecate all synchronous/blocking API in libnm
Note that D-Bus is fundamentally asynchronous. Doing blocking calls
on top of that is wrong, especially for libnm. That is because libnm
also provides a client-side cache of the objects from the D-Bus
interface. This cache must be filled by asynchronous events (of course).
So, making a blocking D-Bus call means to wait for a response and
return it, while queuing everything that happens in between. That
violates the strict ordering that we want to guarantee. It also means,
when the blocking call returns, then the cache is still frozen in the
state as it was before the call, and it will stay so until you iterate
the main context.

Read also [1], for why blocking calls are wrong.

[1] https://smcv.pseudorandom.co.uk/2008/11/nonblocking/

Mark all this API as deprecated. Also, this serves the purpose of
identifying blocking D-Bus calls in libnm.

Note that API like nm_device_wifi_request_scan_options() indeed
does not really need to be in sync with the platform cache. I mean,
it makes a blocking request to start scanning, which does not change
the content of the cache and there wouldn't be a severe problem with
this API. However, it requires as source-argument a NMDeviceWifi, instead
of a plain D-Bus path. If that API would be just a convenience wrapper around
g_dbus_connection_call_sync(), and accept as target a plan D-Bus path,
it would be fine. But having synchronous API that operates on a
NMDeviceWifi is wrong. Hence deprecate it too. Also, it is still wrong
to process any messages from D-Bus (the reply in this case) before other
signals as it destroys the ordering.