CONTRIBUTING: update comment about requiring LGPL-2.1+ license instead of LGPL-2.0+
Historically, libnm claimed to be LGPL-2.0+ licensed. But as keep mixing that
with LGPL-2.1+ code (e.g. from systemd), so probably even the parts that
claimed to be LGPL-2.0+, were not entirely correctly doing so.

Anyway, since we switched to SPDX license identifiers, we now claim
everywhere that the right license is LGPL-2.1+. Update the comment to
reflect that.