all: mark code that is attributed to somebody explicitly
The code contains comments with names in the header of the files.
It's unclear what this exactly means. Does it mean, that the
person wrote the code (Author:)? Does the person hold a copyright?
Is this just a thank-you comment?

Tag the lines with "Attributed-to: ". This still doesn't clarify
the meaning of these comments, but at least it conveys that
it's not clear what there is about this person. It also makes
such lines grep-able (if we ever grep the source code to find
copyright holders).

Basically, tag these comments with an explicit (and unique) marker that
indicates that we don't really know what the comment wants to say.

Of course, in the future we should not add further such comments.